Python

+++ Ejercicios Iniciales/Ejercicio_4.py
@@ -1,5 +1,5 @@
 #3. Programa que pida dos números enteros y realice la suma correspondiente
-var1=int(input("Introduce un número: "))
-var2=int(input("Introduce otro número: "))
+var1=float(input("Introduce un número: "))
+var2=float(input("Introduce otro número: "))
 total=var1+var2
 print("El resultado de la suma es: ", total)
